media-search/requirements.txt
gradio
google-cloud-spanner
vertexai
tqdm
pandas
numpy
httpx